+++ network/src/internal/mod.rs
@@ -19,3 +19,58 @@ pub use connection_handler::*;
 
 pub mod message_handler;
 pub use message_handler::*;
+
+use crate::{external::propagate_transaction, server::Context};
+use snarkos_consensus::{
+    memory_pool::{Entry, MemoryPool},
+    ConsensusParameters,
+    MerkleTreeLedger,
+};
+use snarkos_dpc::base_dpc::{
+    instantiated::{Components, Tx},
+    parameters::PublicParameters,
+};
+use snarkos_errors::network::SendError;
+use snarkos_utilities::bytes::FromBytes;
+
+use std::{net::SocketAddr, sync::Arc};
+use tokio::sync::Mutex;
+
+/// Verify a transaction, add it to the memory pool, propagate it to peers.
+pub async fn process_transaction_internal(
+    context: Arc<Context>,
+    consensus: &ConsensusParameters,
+    parameters: &PublicParameters<Components>,
+    storage: Arc<MerkleTreeLedger>,
+    memory_pool_lock: Arc<Mutex<MemoryPool<Tx>>>,
+    transaction_bytes: Vec<u8>,
+    transaction_sender: SocketAddr,
+) -> Result<(), SendError> {
+    if let Ok(transaction) = Tx::read(&transaction_bytes[..]) {
+        let mut memory_pool = memory_pool_lock.lock().await;
+
+        if !consensus.verify_transaction(parameters, &transaction, &storage)? {
+            error!("Received a transaction that was invalid");
+            return Ok(());
+        }
+
+        if transaction.value_balance.is_negative() {
+            error!("Received a transaction that was a coinbase transaction");
+            return Ok(());
+        }
+
+        let entry = Entry::<Tx> {
+            size: transaction_bytes.len(),
+            transaction,
+        };
+
+        if let Ok(inserted) = memory_pool.insert(&storage, entry) {
+            if inserted.is_some() {
+                info!("Transaction added to memory pool.");
+                propagate_transaction(context.clone(), transaction_bytes, transaction_sender).await?;
+            }
+        }
+    }
+
+    Ok(())
+}
