+++ network/src/server/mod.rs
@@ -19,79 +19,3 @@ pub use context::*;
 
 pub mod server;
 pub use server::*;
-
-use crate::external::message_types::{Block, Transaction};
-
-use snarkos_consensus::{
-    memory_pool::{Entry, MemoryPool},
-    ConsensusParameters,
-    MerkleTreeLedger,
-};
-use snarkos_dpc::base_dpc::{
-    instantiated::{Components, Tx},
-    parameters::PublicParameters,
-};
-use snarkos_errors::network::SendError;
-use snarkos_utilities::bytes::FromBytes;
-
-use std::{net::SocketAddr, sync::Arc};
-use tokio::sync::Mutex;
-
-/// Verify a transaction, add it to the memory pool, propagate it to peers.
-pub async fn process_transaction_internal(
-    context: Arc<Context>,
-    consensus: &ConsensusParameters,
-    parameters: &PublicParameters<Components>,
-    storage: Arc<MerkleTreeLedger>,
-    memory_pool_lock: Arc<Mutex<MemoryPool<Tx>>>,
-    transaction_bytes: Vec<u8>,
-    transaction_sender: SocketAddr,
-) -> Result<(), SendError> {
-    if let Ok(transaction) = Tx::read(&transaction_bytes[..]) {
-        let mut memory_pool = memory_pool_lock.lock().await;
-
-        if !consensus.verify_transaction(parameters, &transaction, &storage)? {
-            error!("Received a transaction that was invalid");
-            return Ok(());
-        }
-
-        if transaction.value_balance.is_negative() {
-            error!("Received a transaction that was a coinbase transaction");
-            return Ok(());
-        }
-
-        let entry = Entry::<Tx> {
-            size: transaction_bytes.len(),
-            transaction,
-        };
-
-        if let Ok(inserted) = memory_pool.insert(&storage, entry) {
-            if inserted.is_some() {
-                debug!("Transaction added to memory pool. Propagating transaction to peers");
-
-                let peer_book = context.peer_book.read().await;
-                let local_address = *context.local_address.read().await;
-                let connections = context.connections.read().await;
-                let mut num_peers = 0;
-
-                for (socket, _) in &peer_book.get_connected() {
-                    if *socket != transaction_sender && *socket != local_address {
-                        if let Some(channel) = connections.get(socket) {
-                            match channel.write(&Transaction::new(transaction_bytes.clone())).await {
-                                Ok(_) => num_peers += 1,
-                                Err(error) => warn!(
-                                    "Failed to propagate transaction to peer {}. (error message: {})",
-                                    channel.address, error
-                                ),
-                            }
-                        }
-                    }
-                }
-
-                debug!("Transaction propagated to {} peers", num_peers);
-            }
-        }
-    }
-
-    Ok(())
-}
